+++ src/main/java/pl/startrader/model/heavenly_body/HeavenlyBody.java
@@ -9,9 +9,9 @@ public abstract class HeavenlyBody {
     public abstract String getName();
     public abstract String getType();
     public abstract Integer getSize();
-    public abstract Integer getLocationX();
-    public abstract Integer getLocationY();
-    public abstract Integer getLocationZ();
+    public abstract Double getLocationX();
+    public abstract Double getLocationY();
+    public abstract Double getLocationZ();
     public abstract List<Resource> getMainResources();
 
     @Override
