+++ src/main/java/pl/startrader/model/heavenly_body/HeavenlyBodyFactory.java
@@ -3,7 +3,7 @@ package pl.startrader.model.heavenly_body;
 public class HeavenlyBodyFactory {
 
     public static HeavenlyBody getHeavenlyBody(String type, String name, Integer size,
-                                               Integer locationX, Integer locationY, Integer locationZ) {
+                                               Double locationX, Double locationY, Double locationZ) {
 
         if("Planet".equalsIgnoreCase(type)) {
             return new Planet(name, size, locationX, locationY, locationZ);
